Refract - split into files
