+++ drafter.gyp
@@ -41,11 +41,16 @@
         "src/RefractAST.h",
         "src/RefractAST.cc",
 
+        # librefract parts - will be separated into other project
+
         "src/refract/Element.h",
         "src/refract/Element.cc",
         "src/refract/Typelist.h",
         "src/refract/VisitableBy.h",
+
         "src/refract/Visitor.h",
+        "src/refract/Visitors.h",
+
         "src/refract/SerializeCompactVisitor.h",
         "src/refract/SerializeCompactVisitor.cc",
         "src/refract/SerializeVisitor.h",
@@ -54,6 +59,15 @@
         "src/refract/ComparableVisitor.cc",
         "src/refract/TypeQueryVisitor.h",
         "src/refract/TypeQueryVisitor.cc",
+        "src/refract/IsExpandableVisitor.h",
+        "src/refract/IsExpandableVisitor.cc",
+        "src/refract/ExpandVisitor.h",
+        "src/refract/ExpandVisitor.cc",
+
+        "src/refract/Registry.h",
+        "src/refract/Registry.cc",
+
+        "src/refract/AppendDecorator.h",
       ],
 
       # FIXME: replace by direct dependecies
