+++ src/refract/Element.h
@@ -18,13 +18,12 @@
 
 
 #include "Exception.h"
-#include "Visitor.h"
-#include "SerializeVisitor.h"
-#include "SerializeCompactVisitor.h"
-#include "ComparableVisitor.h"
-#include "TypeQueryVisitor.h"
 
-#include <iostream>
+#include "Visitor.h"
+
+//#include "AppendDecorator.h"
+
+//#include <iostream>
 
 namespace refract
 {
@@ -38,8 +37,12 @@ namespace refract
     struct ObjectElement;
     struct MemberElement;
 
+    struct ComparableVisitor;
+    struct SerializeVisitor;
+    struct SerializeCompactVisitor;
     struct ExpandVisitor;
     struct IsExpandableVisitor;
+    struct TypeQueryVisitor;
 
     template <typename T> struct ElementTypeSelector;
 
@@ -90,7 +93,8 @@ namespace refract
             SerializeCompactVisitor, 
             ExpandVisitor, 
             IsExpandableVisitor,
-            TypeQueryVisitor>::type Visitors;
+            TypeQueryVisitor
+        >::type Visitors;
 
 
         IElement() : hasContent(false), useCompactContent(false)
@@ -443,396 +447,6 @@ namespace refract
             value.push_back(e);
         }
     };
-
-    class Registry {
-        // FIXME: potentionally dangerous,
-        // if element is deleted and not removed from registry
-        // solution: std::shared_ptr<> || std::weak_ptr<>
-        typedef std::map<std::string, IElement*> Map;
-        Map registrated;
-
-        std::string getElementId(IElement* element) 
-        {
-            IElement::MemberElementCollection::const_iterator it = element->meta.find("id");
-            if (it == element->meta.end()) {
-                throw LogicError("Element has no ID");
-            }
-
-            SerializeCompactVisitor v;
-            (*it)->value.second->content(v);
-            // FIXME: it is really ugly,
-            // make something like is_a()
-            return v.value().str;
-        }
-
-    public:
-
-        IElement* find(const std::string& name)
-        {
-            Map::iterator i = registrated.find(name);
-            if(i == registrated.end()) {
-                return NULL;
-            }
-            return i->second;
-        }
-
-        bool add(IElement* element) 
-        {
-            IElement::MemberElementCollection::const_iterator it = element->meta.find("id");
-            if (it == element->meta.end()) {
-                throw LogicError("Element has no ID");
-            }
-
-            std::string id = getElementId(element);
-            //std::cout << "ID: >" << id << "<"<< std::endl;
-
-            if(isReserved(id)) {
-                throw LogicError("You can not registrate basic element");
-            }
-
-            if (find(id)) {
-                // there is already already element with given name
-                return false;
-            }
-
-            registrated[id] = element;
-            return true;
-        }
-
-        bool remove(const std::string& name) {
-            Map::iterator i = registrated.find(name);
-            if(i == registrated.end()) {
-                return false;
-            }
-            registrated.erase(i);
-            return true;
-        }
-
-        void clearAll(bool releaseElements = false) {
-            if(releaseElements) {
-                for (Map::iterator i = registrated.begin() ; i != registrated.end() ; ++i) {
-                    delete i->second;
-                }
-            }
-            registrated.clear();
-        }
-
-    };
-
-    // FIXME: remove global variable!!
-    extern Registry DSRegistry;
-
-    template <typename T, typename V = typename T::ValueType>
-    struct AppendDecorator {
-        typedef T ElementType;
-        typedef V ValueType;
-        ElementType*& element;
-        AppendDecorator(ElementType*& e) : element(e) 
-        {
-        }
-
-        void operator()(const V& value) 
-        {
-            //throw std::logic_error("Can not append to primitive type");
-            // FIXME: snowcrash warn about "Primitive type can not have member"
-            // but in real it create "empty" member
-            //
-            // solution for now: silently ignore
-        }
-    };
-
-    template <typename T>
-    struct AppendDecorator<T, std::string> {
-        typedef T ElementType;
-        typedef typename T::ValueType ValueType;
-        ElementType*& element;
-
-        AppendDecorator(ElementType*& e) : element(e) 
-        {
-        }
-
-        void operator()(const std::string& value) {
-            if(!value.empty()) {
-                element->value.append(value);
-            }
-        }
-    };
-
-    template <typename T>
-    struct AppendDecorator<T, std::vector<refract::IElement*> > {
-        typedef T ElementType;
-        typedef typename T::ValueType ValueType;
-        ElementType*& element;
-
-        AppendDecorator(ElementType*& e) : element(e) 
-        {
-        }
-
-        void operator()(const ValueType& value)
-        {
-            std::copy(value.begin(), value.end(), std::back_inserter(element->value));
-            if(!value.empty()) {
-                element->hasContent = true;
-            }
-        }
-    };
-
-    struct IsExpandableVisitor : public IVisitor {
-
-        struct CheckElement {
-            bool checkElement(const IElement* e) const {
-                std::string type;
-                if(e) {
-                    type = e->element();
-                }
-                return !isReserved(type) || type == "ref";
-            }
-        };
-
-        template <typename T, typename V = typename T::ValueType>
-        struct IsExpandable : public CheckElement {
-            bool operator()(const T* e) const {
-                if(checkElement(e)) {
-                    return true;
-                }
-                return false;
-            }
-        };
-
-        template <typename T>
-        struct IsExpandable<T, MemberElement::ValueType> : public CheckElement {
-            bool operator()(const T* e) const {
-                //std::cout << __PRETTY_FUNCTION__ << e->element() << std::endl;
-
-                if(checkElement(e)) {
-                    return true;
-                }
-
-                if(e->value.first) {
-                    IsExpandableVisitor v;
-                    e->value.first->content(v);
-                    if (v.result) {
-                        return true;
-                    }
-                }
-
-                if(e->value.second) {
-                    IsExpandableVisitor v;
-                    e->value.second->content(v);
-                    if (v.result) {
-                        return true;
-                    }
-                }
-
-
-                return false;
-            }
-        };
-
-        template <typename T>
-        struct IsExpandable<T, std::vector<IElement*> > : public CheckElement {
-            bool operator()(const T* e) const {
-                //std::cout << __PRETTY_FUNCTION__ << e->element() << std::endl;
-
-                if(checkElement(e)) {
-                    return true;
-                }
-
-                for (std::vector<IElement*>::const_iterator i = e->value.begin() ; i != e->value.end() ; ++i ) {
-                    IsExpandableVisitor v;
-                    (*i)->content(v);
-                    if (v.result) {
-                        return true;
-                    }
-                }
-
-                return false;
-            }
-        };
-
-        bool result;
-        IsExpandableVisitor() : result(false) {}
-
-        template<typename T>
-        void visit(const T& e) {
-            //std::cout << __PRETTY_FUNCTION__ << std::endl;
-            result = IsExpandable<T>()(&e);
-        }
-
-        bool get() const {
-            return result;
-        }
-    };
-
-
-    struct ExpandVisitor : IVisitor {
-
-        IElement* result;
-
-        static IElement* expandOrClone(const IElement* e) {
-            IElement* result = NULL;
-            if(!e) {
-                return result;
-            }
-
-            ExpandVisitor expander;
-            e->content(expander);
-            result = expander.get();
-
-            if(!result) {
-                result = e->clone();
-            }
-
-            return result;
-        }
-
-        ExpandVisitor() : result(NULL) {};
-
-        void visit(const IElement& e) {
-            //std::cout << __PRETTY_FUNCTION__ <<  std::endl;
-
-            IsExpandableVisitor isExpandable;
-            e.content(isExpandable);
-            if (!isExpandable.get()) {
-                return;
-            }
-
-            ExpandVisitor expander;
-            e.content(expander);
-
-            result = expander.get();
-        }
-
-        void visit(const MemberElement& e) {
-            //std::cout << __PRETTY_FUNCTION__ <<  std::endl;
-
-            IsExpandableVisitor isExpandable;
-            e.content(isExpandable);
-            if (!isExpandable.get()) {
-                return;
-            }
-
-            MemberElement* expanded = static_cast<MemberElement*>(e.clone(IElement::cAll ^ IElement::cValue));
-
-            expanded->set(expandOrClone(e.value.first), expandOrClone(e.value.second));
-
-            result = expanded;
-        }
-
-        void visit(const ObjectElement& e) {
-            //std::cout << __PRETTY_FUNCTION__ <<  std::endl;
-
-            std::string en = e.element();
-
-            // FIXME: refactoring - split into method
-            if(!isReserved(en)) { // handle direct inheritance
-                refract::ObjectElement* o = new refract::ObjectElement;
-                o->element("extend");
-
-                IElement::MemberElementCollection::const_iterator name = e.meta.find("id");
-                if (name != e.meta.end() && (*name)->value.second) {
-                    o->meta["id"] = (*name)->value.second->clone();
-                }
-
-                // go as deep as possible in inheritance tree
-                for (const IElement* parent = refract::DSRegistry.find(en)
-                    ; parent && !isReserved(en)
-                    ; en = parent->element(), parent = refract::DSRegistry.find(en) ) {
-
-                    // FIXME: while clone original element w/o meta - we lose `description`
-                    // must be fixed in spec
-                    IElement* clone = parent->clone(IElement::cAll ^ IElement::cMeta);
-                    clone->meta["ref"] = IElement::Create(en);
-                    o->push_back(clone);
-                }
-
-                ObjectElement* origin = new ObjectElement; // warapper for original object
-                for (ObjectElement::ValueType::const_iterator it = e.value.begin() ; it != e.value.end() ; ++it) {
-                    origin->push_back(expandOrClone(*it));
-                }
-                o->push_back(origin);
-
-                result = o;
-                return;
-            }
-            else if (en == "ref") {
-                //refract::ObjectElement* o = new refract::ObjectElement;
-                refract::ObjectElement* o = new refract::ObjectElement;
-
-                for (ObjectElement::ValueType::const_iterator it = e.value.begin()
-                    ; it != e.value.end()
-                    ; ++it ) {
-
-                    ComparableVisitor href("href", ComparableVisitor::key);
-                    (*it)->content(href);
-
-                    if(href.get()) { // key was recognized - it is save to cast to MemberElement
-                        MemberElement* m = static_cast<MemberElement*>(*it);
-                        if(m->value.second) {
-                            TypeQueryVisitor tq;
-                            if(StringElement* value = tq.as<StringElement>(m->value.second)) {
-                                en = value->value;
-
-                                // go as deep as possible in inheritance tree
-                                for (const IElement* parent = refract::DSRegistry.find(en)
-                                   ; parent && !isReserved(en)
-                                   ; en = parent->element(), parent = refract::DSRegistry.find(en) ) {
-
-                                   // FIXME: while clone original element w/o meta - we lose `description`
-                                   // must be fixed in spec
-                                   IElement* clone = parent->clone(IElement::cAll ^ IElement::cMeta);
-                                   clone->meta["ref"] = IElement::Create(en);
-                                   o->push_back(clone);
-                                }
-                            }
-                        }
-                    }
-                }
-
-                result = o;
-                return;
-            }
-
-            // handle expandable members
-            IsExpandableVisitor isExpandable;
-            e.content(isExpandable);
-            if (!isExpandable.get()) {
-               //std::cout << "not expandable - leave" << std::endl;
-                return;
-            }
-
-            // do not clone value, we have to do it one by one because some of them must be expanded
-            refract::ObjectElement* expanded = static_cast<ObjectElement*>(e.clone(IElement::cAll ^ IElement::cValue));
-
-            std::vector<IElement*> members;
-            bool hasRef = false;
-
-            for (ObjectElement::ValueType::const_iterator it = e.value.begin() ; it != e.value.end() ; ++it) {
-                if((*it)->element() == "ref") {
-                   hasRef = true; 
-                }
-
-                members.push_back(expandOrClone(*it));
-            }
-
-            expanded->set(members);
-
-            if (hasRef) {
-                ObjectElement* extend = new ObjectElement;
-                extend->element("extend");
-                extend->push_back(expanded);
-                expanded = extend;
-            }
-
-            result = expanded;
-        }
-
-        IElement* get() const {
-            return result;
-        }
-    };
-
-
-}; // namespace refract
+};
 
 #endif // #ifndef _REFRACT_ELEMENT_H_
