+++ src/refract/Registry.h
@@ -1,0 +1,44 @@
+//
+//  refract/Registry.h
+//  librefract
+//
+//  Created by Jiri Kratochvil on 21/05/15.
+//  Copyright (c) 2015 Apiary Inc. All rights reserved.
+//
+#ifndef _REFRACT_REGISTRY_H_
+#define _REFRACT_REGISTRY_H_
+
+#include <map>
+#include <string>
+
+namespace refract
+{
+
+    // Forward declarations of IElement
+    struct IElement;
+
+    class Registry {
+        // FIXME: potentionally dangerous,
+        // if element is deleted and not removed from registry
+        // solution: std::shared_ptr<> || std::weak_ptr<>
+        typedef std::map<std::string, IElement*> Map;
+        Map registrated;
+
+        std::string getElementId(IElement* element);
+
+    public:
+
+        IElement* find(const std::string& name);
+
+        bool add(IElement* element) ;
+        bool remove(const std::string& name);
+        void clearAll(bool releaseElements = false);
+    };
+
+    // FIXME: remove global variable!!
+    extern Registry DSRegistry;
+
+
+}; // namespace refract
+
+#endif // #ifndef _REFRACT_REGISTRY_H_
