+++ src/refract/Visitors.h
@@ -1,0 +1,18 @@
+//
+//  refract/Visitors.h
+//  librefract
+//
+//  Created by Jiri Kratochvil on 18/05/15.
+//  Copyright (c) 2015 Apiary Inc. All rights reserved.
+//
+#ifndef _REFRACT_VISITORS_H_
+#define _REFRACT_VISITORS_H_
+
+#include "ComparableVisitor.h"
+#include "SerializeVisitor.h"
+#include "SerializeCompactVisitor.h"
+#include "IsExpandableVisitor.h"
+#include "ExpandVisitor.h"
+#include "TypeQueryVisitor.h"
+
+#endif // #ifndef _REFRACT_VISITORS_H_
